change(pallets/communities): split Pallet internal functions into several modules
