+++ pallets/communities/src/functions/mod.rs
@@ -1,0 +1,7 @@
+pub(self) use crate::*;
+pub(self) use frame_support::pallet_prelude::*;
+pub(self) use types::*;
+
+mod getters;
+mod membership;
+mod registration;
